Refactor Search component, minor styling tweaks
